requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.7
numpy>=1.24.0
openai-whisper>=20231117
Pillow>=10.0.0
beautifulsoup4>=4.12.0
fastapi>=0.100.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
scenedetect>=0.6.0
selenium>=4.15.0
google-generativeai>=0.3.2
redis>=4.5.0
rq>=1.15.0
rq-dashboard>=0.6.1 